requirements.txt
gspread==5.10.0
google-auth==2.20.0
pandas
streamlit